Refactor context, mint page doesnt error
